Added grpc-web support checks and exceptions

+++ packages/eventstore_client/lib/src/core/helpers.dart
@@ -114,7 +114,7 @@ extension DurationX on Duration {
   bool get isNone => this == Defaults.NoneDuration;
 
   /// See [Defaults.InfiniteDuration]
-  bool get isInfinite => inMilliseconds == Max.toInt();
+  bool get isInfinite => inMilliseconds == Max;
 }
 
 extension ListX<T> on List<T> {
