+++ packages/eventstore_client/lib/src/persistent/persistent_subscription_settings.dart
@@ -21,8 +21,7 @@ class PersistentSubscriptionSettings {
     this.messageTimeout = const Duration(seconds: 30),
     this.consumerStrategy = SystemConsumerStrategies.roundRobin,
   }) {
-    if (messageTimeout < Duration.zero ||
-        messageTimeout.inMilliseconds > Max.toInt()) {
+    if (messageTimeout < Duration.zero || messageTimeout.inMilliseconds > Max) {
       throw ArgumentError.value(
         messageTimeout,
         'messageTimeout',
@@ -31,7 +30,7 @@ class PersistentSubscriptionSettings {
     }
 
     if (checkPointAfter < Duration.zero ||
-        checkPointAfter.inMilliseconds > Max.toInt()) {
+        checkPointAfter.inMilliseconds > Max) {
       throw ArgumentError.value(
         checkPointAfter,
         'checkPointAfter',
